lab5-kyber-crystal-echo/requirements.txt
# Lab5: KYB-02 Crystal Echo - Dependencies
# ==========================================

# Core PQC Library
liboqs-python>=0.10.0

# Testing Framework
pytest>=7.4.0
pytest-timeout>=2.2.0
pytest-cov>=4.1.0

# Fuzzing Tools
atheris>=2.3.0
python-afl>=0.7.3

# Cryptographic Utilities
pycryptodome>=3.19.0
cryptography>=41.0.0

# Performance & Timing Analysis
# Se comenta porque es una herramienta de sistema, no de python
# hyperfine>=1.16.0 

# Utilities
numpy>=1.24.0
rich>=13.5.0
